requirements.txt
google-api-python-client
boto3
apache-airflow-providers-amazon
minio
apache-airflow-providers-apache-spark
oauth2client
